+++ mirumon/application/devices/auth_service.py
@@ -7,7 +7,7 @@ from passlib.context import CryptContext
 from pydantic import BaseModel, SecretStr
 
 from mirumon.domain.devices.entities import Device, DeviceID
-from mirumon.settings.environments.base import AppSettings
+from mirumon.settings.environments.app import AppSettings
 
 
 class MetaJWT(BaseModel):
