+++ mirumon/application/devices/devices_service.py
@@ -1,7 +1,7 @@
 from mirumon.application.devices.device_repo import DevicesRepository
 from mirumon.application.events.events_repo import EventsRepository
 from mirumon.domain.devices.entities import Device
-from mirumon.settings.environments.base import AppSettings
+from mirumon.settings.environments.app import AppSettings
 
 
 class DevicesService:
