+++ mirumon/infra/api/dependencies/users/users_services.py
@@ -5,7 +5,7 @@ from mirumon.application.users.users_repo import UsersRepository
 from mirumon.application.users.users_service import UsersService
 from mirumon.infra.api.dependencies.repositories import get_repository
 from mirumon.settings.config import get_app_settings
-from mirumon.settings.environments.base import AppSettings
+from mirumon.settings.environments.app import AppSettings
 
 
 def _get_users_service(
