+++ mirumon/infra/api/devices/devices_router.py
@@ -2,42 +2,42 @@ from fastapi import APIRouter, Depends
 
 from mirumon.domain.users.scopes import DevicesScopes
 from mirumon.infra.api.dependencies.users.permissions import check_user_scopes
-from mirumon.infra.api.endpoints.devices.http import (
-    auth,
-    detail,
-    execute,
-    info,
-    list,
-    shutdown,
+from mirumon.infra.api.devices.http_endpoints import (
+    detail_controller,
+    devices_controller,
+    execute_controller,
+    info_controller,
+    list_controller,
+    shutdown_controller,
 )
-from mirumon.infra.api.endpoints.devices.ws import connect
+from mirumon.infra.api.devices.ws_endpoints import connect_controller
 
 router = APIRouter(tags=["Devices"])
 
 
 # http
-router.include_router(auth.router)
+router.include_router(devices_controller.router)
 router.include_router(
-    detail.router,
+    detail_controller.router,
     dependencies=[Depends(check_user_scopes([DevicesScopes.read]))],
 )
 router.include_router(
-    list.router,
+    list_controller.router,
     dependencies=[Depends(check_user_scopes([DevicesScopes.read]))],
 )
 
 router.include_router(
-    info.router,
+    info_controller.router,
     dependencies=[Depends(check_user_scopes([DevicesScopes.read]))],
 )
 router.include_router(
-    shutdown.router,
+    shutdown_controller.router,
     dependencies=[Depends(check_user_scopes([DevicesScopes.write]))],
 )
 router.include_router(
-    execute.router,
+    execute_controller.router,
     dependencies=[Depends(check_user_scopes([DevicesScopes.write]))],
 )
 
 # websockets
-router.include_router(connect.router)
+router.include_router(connect_controller.router)
