+++ mirumon/infra/api/devices/http_endpoints/execute_controller.py
@@ -6,7 +6,7 @@ from mirumon.application.events.events_service import EventsService
 from mirumon.application.events.models import EventParams, EventTypes
 from mirumon.domain.devices.entities import DeviceID
 from mirumon.infra.api.dependencies.services import get_service
-from mirumon.infra.api.endpoints.devices.http.models.execute import ExecuteCommandParams
+from mirumon.infra.api.devices.http_endpoints.models.execute import ExecuteCommandParams
 
 DEVICE_UNAVAILABLE_ERROR = HTTPException(
     status_code=status.HTTP_503_SERVICE_UNAVAILABLE, detail="device unavailable"
