+++ mirumon/infra/api/devices/http_endpoints/models/detail.py
