+++ mirumon/infra/api/devices/http_endpoints/models/hardware.py
