+++ mirumon/infra/api/devices/http_endpoints/shutdown_controller.py
