+++ mirumon/infra/api/routers.py
@@ -1,10 +1,10 @@
 from fastapi import APIRouter
 
-from mirumon.infra.api.endpoints.devices.router import router as devices_router
-from mirumon.infra.api.endpoints.users.router import router as users_router
+from mirumon.infra.api.devices import devices_router
+from mirumon.infra.api.users import users_router
 
 router = APIRouter()
 
 
-router.include_router(users_router)
-router.include_router(devices_router)
+router.include_router(users_router.router)
+router.include_router(devices_router.router)
