+++ mirumon/infra/api/users/users_router.py
@@ -1,0 +1,13 @@
+from fastapi import APIRouter, Depends
+
+from mirumon.domain.users.scopes import UsersScopes
+from mirumon.infra.api.dependencies.users.permissions import check_user_scopes
+from mirumon.infra.api.users.http_endpoints import tokens_controllers, users_controllers
+
+router = APIRouter(tags=["Users"])
+
+router.include_router(tokens_controllers.router)
+router.include_router(
+    users_controllers.router,
+    dependencies=[Depends(check_user_scopes([UsersScopes.write]))],
+)
