+++ mirumon/settings/environments/development.py
@@ -1,6 +1,6 @@
 from pydantic import AnyUrl, PostgresDsn, SecretStr
 
-from mirumon.settings.environments.base import AppSettings
+from mirumon.settings.environments.app import AppSettings
 
 
 class DevAppSettings(AppSettings):
