+++ tests/integration/infra/api/devices/http_endpoints/test_devices_controller.py
