+++ tests/integration/infra/api/devices/http_endpoints/test_execute_controller.py
