+++ tests/integration/infra/api/devices/http_endpoints/test_info_controller.py
