+++ tests/integration/infra/api/devices/http_endpoints/test_shutdown_controller.py
